+++ src/app/index.tsx
@@ -2,6 +2,7 @@ import { Container } from "@/components/layout/container";
 import { useOkamiToast } from "@/components/okami-toast";
 import { Center } from "@/components/ui/center";
 import { Spinner } from "@/components/ui/spinner";
+
 import { isAuthAction } from "@/store/auth";
 import { Redirect } from "expo-router";
 import { useAtomValue } from "jotai";
